+++ src/schedule/schedule.service.ts
@@ -11,8 +11,12 @@ import { UserService } from '../user/user.service';
 import { Tag } from '../tag/tag.entity';
 import { ChannelService } from '../channel/channel.service';
 import { WebClient } from '@slack/web-api';
-import { KnownBlock } from '@slack/web-api';
 import { randomUUID } from 'crypto';
+import {
+  buildRecurringCreationBlocks,
+  buildRecurringDeleteBlocks,
+  buildRecurringUpdateBlocks,
+} from './view/schedule-recurring.view';
 import { RRule, Weekday } from 'rrule';
 import { ScheduleNotificationService } from './schedule-notification.service';
 import { CreateScheduleDto, UpdateScheduleDto } from './dto/schedule.dto';
@@ -1049,380 +1053,6 @@ async function runInChunks<T, R>(
     }
   }
   return allResults;
-}
-
-function parseLocationParts(location: string): {
-  room: string;
-  professor: string | null;
-} {
-  const slashIdx = location.indexOf('/');
-  if (slashIdx === -1) return { room: location.trim(), professor: null };
-  return {
-    room: location.slice(0, slashIdx).trim(),
-    professor: location.slice(slashIdx + 1).trim() || null,
-  };
-}
-
-function buildRecurringDeleteBlocks(
-  scheduleName: string,
-  group: RecurrenceGroup,
-  scope: 'all' | 'future',
-  filterOriginal: boolean,
-  deletedCount: number,
-  totalCount: number,
-  executorDisplay: string,
-  writerDisplay: string,
-): KnownBlock[] {
-  const DAY_LABELS = ['일', '월', '화', '수', '목', '금', '토'];
-  const fmtDays = (days: number[] | null) =>
-    days && days.length > 0
-      ? [...days]
-          .sort((a, b) => a - b)
-          .map((d) => DAY_LABELS[d])
-          .join('·')
-      : '미지정';
-  const fmtTime = (s: string, e: string) => `${s} - ${e}`;
-  const fmtPeriod = (s: string, e: string) => `${s} - ${e}`;
-
-  const scopeText = scope === 'all' ? '전체' : '오늘 이후';
-  const filterText = filterOriginal ? '원본만' : '전체';
-  const statusText =
-    deletedCount < totalCount
-      ? `⚠️ ${deletedCount}/${totalCount}개 삭제 완료`
-      : `총 ${deletedCount}개 삭제`;
-
-  const { room: delRoom, professor: delProfessor } = parseLocationParts(
-    group.location ?? '',
-  );
-
-  return [
-    {
-      type: 'header',
-      text: {
-        type: 'plain_text',
-        text: `🗑️ [${scheduleName}] 반복 일정 삭제 안내`,
-        emoji: true,
-      },
-    },
-    { type: 'divider' },
-    {
-      type: 'section',
-      fields: [
-        { type: 'mrkdwn', text: `📌 *일정 제목*\n*${group.title}*` },
-        {
-          type: 'mrkdwn',
-          text: `🕐 *시간*\n${fmtTime(group.startTime, group.endTime)}`,
-        },
-      ],
-    },
-    {
-      type: 'section',
-      fields: [
-        {
-          type: 'mrkdwn',
-          text: `🗓️ *기간*\n${fmtPeriod(group.startDate, group.endDate)}`,
-        },
-        {
-          type: 'mrkdwn',
-          text: `📅 *요일*\n${fmtDays(group.daysOfWeek)}`,
-        },
-      ],
-    },
-    delProfessor
-      ? {
-          type: 'section',
-          fields: [
-            {
-              type: 'mrkdwn',
-              text: `📍 *장소*\n${delRoom || '_미지정_'}`,
-            },
-            { type: 'mrkdwn', text: `👨‍🏫 *교수*\n${delProfessor}` },
-          ],
-        }
-      : {
-          type: 'section',
-          text: {
-            type: 'mrkdwn',
-            text: `📍 *장소*\n${delRoom || '_미지정_'}`,
-          },
-        },
-    {
-      type: 'section',
-      fields: [
-        { type: 'mrkdwn', text: `👤 *실행자*\n${executorDisplay}` },
-        { type: 'mrkdwn', text: `👥 *담당자*\n${writerDisplay}` },
-      ],
-    },
-    { type: 'divider' },
-    {
-      type: 'context',
-      elements: [
-        {
-          type: 'mrkdwn',
-          text: `🔍 *삭제 범위:* ${scopeText} (${filterText}) | 📊 *처리 결과:* ${statusText} | *Bannote Bot*`,
-        },
-      ],
-    },
-  ] as unknown as KnownBlock[];
-}
-
-function buildRecurringUpdateBlocks(
-  scheduleName: string,
-  group: RecurrenceGroup,
-  dto: UpdateRecurringEventsDto,
-  scope: 'all' | 'future',
-  updatedCount: number,
-  totalCount: number,
-  executorDisplay: string,
-  writerDisplay: string,
-): KnownBlock[] {
-  const DAY_LABELS = ['일', '월', '화', '수', '목', '금', '토'];
-  const fmtDays = (days: number[] | null) =>
-    days && days.length > 0
-      ? [...days]
-          .sort((a, b) => a - b)
-          .map((d) => DAY_LABELS[d])
-          .join('·')
-      : '미지정';
-  const fmtTime = (s: string | null, e: string | null) =>
-    s && e ? `${s} - ${e}` : '미지정';
-  const fmtPeriod = (s: string | null, e: string | null) =>
-    s && e ? `${s} - ${e}` : '미지정';
-
-  // 제목
-  const titleChanged = dto.title !== undefined && dto.title !== group.title;
-  const titleText = titleChanged
-    ? `📌 *일정 제목* ✏️\n~${group.title}~ → *${dto.title}*`
-    : `📌 *일정 제목*\n*${group.title}*`;
-
-  // 시간
-  const newStart = dto.startTime ?? group.startTime;
-  const newEnd = dto.endTime ?? group.endTime;
-  const timeChanged =
-    (dto.startTime !== undefined && dto.startTime !== group.startTime) ||
-    (dto.endTime !== undefined && dto.endTime !== group.endTime);
-  const timeText = timeChanged
-    ? `🕐 *시간* ✏️\n~${fmtTime(group.startTime, group.endTime)}~ → *${fmtTime(newStart, newEnd)}*`
-    : `🕐 *시간*\n${fmtTime(group.startTime, group.endTime)}`;
-
-  // 요일
-  const daysChanged =
-    dto.daysOfWeek !== undefined &&
-    fmtDays(dto.daysOfWeek) !== fmtDays(group.daysOfWeek);
-  const daysText = daysChanged
-    ? `📅 *요일* ✏️\n~${fmtDays(group.daysOfWeek)}~ → *${fmtDays(dto.daysOfWeek!)}*`
-    : `📅 *요일*\n${fmtDays(group.daysOfWeek)}`;
-
-  // 기간
-  const newStartDate = dto.startDate ?? group.startDate;
-  const newEndDate = dto.endDate ?? group.endDate;
-  const periodChanged =
-    (dto.startDate !== undefined && dto.startDate !== group.startDate) ||
-    (dto.endDate !== undefined && dto.endDate !== group.endDate);
-  const periodText = periodChanged
-    ? `🗓️ *기간* ✏️\n~${fmtPeriod(group.startDate, group.endDate)}~ \n→ *${fmtPeriod(newStartDate, newEndDate)}*`
-    : `🗓️ *기간*\n${fmtPeriod(group.startDate, group.endDate)}`;
-
-  // 장소 / 교수
-  const locationChanged =
-    dto.location !== undefined && dto.location !== group.location;
-  const afterLocation = dto.location ?? group.location ?? '';
-  const { room: afterRoom, professor: afterProfessor } =
-    parseLocationParts(afterLocation);
-
-  let roomText: string;
-  let professorText: string | null = null;
-
-  if (locationChanged) {
-    const { room: beforeRoom, professor: beforeProfessor } = parseLocationParts(
-      group.location ?? '',
-    );
-
-    roomText =
-      beforeRoom !== afterRoom
-        ? `📍 *장소* ✏️\n~${beforeRoom || '미지정'}~ → *${afterRoom || '미지정'}*`
-        : `📍 *장소*\n${afterRoom || '_미지정_'}`;
-
-    if (beforeProfessor !== afterProfessor) {
-      professorText = `👨‍🏫 *교수* ✏️\n~${beforeProfessor || '미지정'}~ → *${afterProfessor || '미지정'}*`;
-    } else if (afterProfessor) {
-      professorText = `👨‍🏫 *교수*\n${afterProfessor}`;
-    }
-  } else {
-    roomText = `📍 *장소*\n${afterRoom || '_미지정_'}`;
-    if (afterProfessor) {
-      professorText = `👨‍🏫 *교수*\n${afterProfessor}`;
-    }
-  }
-
-  // 처리 결과
-  const scopeText = scope === 'all' ? '전체' : '오늘 이후';
-  const statusText =
-    updatedCount < totalCount
-      ? `⚠️ ${updatedCount}/${totalCount}개 수정 완료`
-      : `총 ${updatedCount}개 수정`;
-
-  return [
-    {
-      type: 'header',
-      text: {
-        type: 'plain_text',
-        text: `🔄 [${scheduleName}] 반복 일정 수정 안내`,
-        emoji: true,
-      },
-    },
-    { type: 'divider' },
-    {
-      type: 'section',
-      fields: [
-        { type: 'mrkdwn', text: titleText },
-        { type: 'mrkdwn', text: timeText },
-      ],
-    },
-    {
-      type: 'section',
-      fields: [
-        { type: 'mrkdwn', text: periodText },
-        { type: 'mrkdwn', text: daysText },
-      ],
-    },
-    professorText
-      ? {
-          type: 'section',
-          fields: [
-            { type: 'mrkdwn', text: roomText },
-            { type: 'mrkdwn', text: professorText },
-          ],
-        }
-      : {
-          type: 'section',
-          text: { type: 'mrkdwn', text: roomText },
-        },
-    {
-      type: 'section',
-      fields: [
-        { type: 'mrkdwn', text: `👤 *실행자*\n${executorDisplay}` },
-        { type: 'mrkdwn', text: `👥 *담당자*\n${writerDisplay}` },
-      ],
-    },
-    { type: 'divider' },
-    {
-      type: 'context',
-      elements: [
-        {
-          type: 'mrkdwn',
-          text: `🔍 *수정 범위:* ${scopeText} | 📊 *처리 결과:* ${statusText} | *Bannote Bot*`,
-        },
-      ],
-    },
-  ] as unknown as KnownBlock[];
-}
-
-// Slack 알림 블록 빌더
-function buildRecurringCreationBlocks(
-  scheduleName: string,
-  dto: CreateRecurringEventsDto,
-  totalCount: number,
-  successCount: number,
-  executorDisplay: string,
-  writerDisplay: string,
-): KnownBlock[] {
-  const DAY_LABELS = ['일', '월', '화', '수', '목', '금', '토'];
-  const recurrenceLabel: Record<RecurrenceType, string> = {
-    weekly: '매주',
-    biweekly: '격주',
-    monthly: '매월',
-  };
-  const fmtDays = (days: number[] | undefined) =>
-    days && days.length > 0
-      ? [...days]
-          .sort((a, b) => a - b)
-          .map((d) => DAY_LABELS[d])
-          .join('·')
-      : '미지정';
-  const fmtTime = (s: string, e: string) => `${s} - ${e}`;
-  const fmtPeriod = (s: string, e: string) => `${s} - ${e}`;
-
-  const recurrenceText =
-    dto.recurrenceType !== 'monthly' && dto.daysOfWeek?.length
-      ? `${recurrenceLabel[dto.recurrenceType]} ${fmtDays(dto.daysOfWeek)}요일`
-      : recurrenceLabel[dto.recurrenceType];
-
-  const statusText =
-    successCount < totalCount
-      ? `⚠️ ${successCount}/${totalCount}개 생성 완료`
-      : `총 ${successCount}개 생성`;
-
-  const { room: crtRoom, professor: crtProfessor } = parseLocationParts(
-    dto.location ?? '',
-  );
-
-  return [
-    {
-      type: 'header',
-      text: {
-        type: 'plain_text',
-        text: `✨ [${scheduleName}] 반복 일정 추가 안내`,
-        emoji: true,
-      },
-    },
-    { type: 'divider' },
-    {
-      type: 'section',
-      fields: [
-        { type: 'mrkdwn', text: `📌 *일정 제목*\n*${dto.title}*` },
-        {
-          type: 'mrkdwn',
-          text: `🕐 *시간*\n${fmtTime(dto.startTime, dto.endTime)}`,
-        },
-      ],
-    },
-    {
-      type: 'section',
-      fields: [
-        {
-          type: 'mrkdwn',
-          text: `🗓️ *기간*\n${fmtPeriod(dto.startDate, dto.endDate)}`,
-        },
-        {
-          type: 'mrkdwn',
-          text: `📅 *요일*\n${fmtDays(dto.daysOfWeek)}`,
-        },
-      ],
-    },
-    crtProfessor
-      ? {
-          type: 'section',
-          fields: [
-            { type: 'mrkdwn', text: `📍 *장소*\n${crtRoom || '_미지정_'}` },
-            { type: 'mrkdwn', text: `👨‍🏫 *교수*\n${crtProfessor}` },
-          ],
-        }
-      : {
-          type: 'section',
-          text: {
-            type: 'mrkdwn',
-            text: `📍 *장소*\n${crtRoom || '_미지정_'}`,
-          },
-        },
-    {
-      type: 'section',
-      fields: [
-        { type: 'mrkdwn', text: `👤 *실행자*\n${executorDisplay}` },
-        { type: 'mrkdwn', text: `👥 *담당자*\n${writerDisplay}` },
-      ],
-    },
-    { type: 'divider' },
-    {
-      type: 'context',
-      elements: [
-        {
-          type: 'mrkdwn',
-          text: `🔁 *반복:* ${recurrenceText} | 📊 *처리 결과:* ${statusText} | *Bannote Bot*`,
-        },
-      ],
-    },
-  ] as unknown as KnownBlock[];
 }
 
 // ========== 원본 이벤트 판별 헬퍼 ==========
